+++ src/dotenv.cpp
@@ -1,9 +1,32 @@
 
 #include <extras/dotenv.hpp>
 #include <extras/strings.hpp>
+#include <iomanip>
 #include <iostream>
 
 namespace extras {
+
+  std::ostream &operator<<(std::ostream &out, const DotENVLine &obj) {
+    out << obj._key << '=' << obj._value;
+    return out;
+  }
+
+  std::istream &operator>>(std::istream &in, DotENVLine &obj) {
+    std::stringstream key;
+    std::string value;
+    char c = ' ';
+    in >> std::skipws;
+    while (in.good() && c != '=') {
+      in >> c;
+      if (c == '#') break;
+      if (isalnum(c) || c == '_') key << c;
+    }
+    DotENVLineKeyException::assertion(key.str(), __INFO__);
+    obj._key = key.str();
+    getline(in, value);
+    obj._value = value;
+    return in;
+  }
 
   std::ostream &operator<<(std::ostream &out, const DotENV &obj) {
     for (auto entry : obj._map) {
@@ -16,9 +39,15 @@ namespace extras {
     std::string line;
     while (in.good()) {
       std::getline(in, line);
+      if (line[0] == '#' || line[0] == ' ' || line[0] == '\n') {
+        continue;
+      }
       if (line.empty()) break;
-      auto parts = split(line, '=');
-      obj.put(parts[0], (parts.size() > 1 ? parts[1] : ""));
+      std::stringstream ss;
+      ss << line;
+      DotENVLine dotENVLine;
+      ss >> dotENVLine;
+      obj.put(dotENVLine);
     }
     return in;
   }